Write configured-int lookup tables from sources of the same dtype
``dest[index] = source`` checks that the two dtypes match, unlike slice
assignment, so an int64 ``arange`` or ``unique`` inverse can no longer be
written into a table that now carries the configured int dtype. The
Friedel-mate maps, the grid-cell lookup, the spherical-harmonic row map
and the atom remap used by hydrogen generation build their sources in the
table's dtype instead.

The Legendre shell kernel TORCH_CHECKs int64 shell labels and offsets, so
those two tensors keep int64 and their markers say why.

+++ torchref/experimental/alignment/sh.py
@@ -138,7 +138,7 @@ def _bar_legendre_recurrence(
         rows = keep_l.to(device=device, dtype=get_int_dtype())
     # l -> its position in the output, or -1 when it is not kept.
     where = torch.full((L,), -1, dtype=get_int_dtype(), device=device)
-    where[rows] = torch.arange(rows.numel(), device=device)
+    where[rows] = torch.arange(rows.numel(), device=device, dtype=where.dtype)
     where_list = where.tolist()
 
     out = torch.zeros((*batch_shape, rows.numel(), L), dtype=dtype, device=device)
